# Boost Scrivener Links Generation Speed