The four largest changes to the orginal implementation are:

  * Removal of runtime format of pagination keys
  * chained building of pagination list
  * recalculation of the base path
  * combined `reduce` from a `map |> into`

### ORIGINAL BENCH

```
 ## Builder.ScrivenerLinksBench
 only one page                  100000   21.63 µs/op
 page one of two pages           50000   58.48 µs/op
 page three of three pages       50000   58.75 µs/op
 page two of two pages           50000   59.19 µs/op
 page one of three pages         50000   59.22 µs/op
 page two of three pages         10000   102.92 µs/op
```

### AFTER BOOST

```
 ## Builder.ScrivenerLinksBench
 only one page                  100000   13.41 µs/op
 page two of two pages           50000   41.97 µs/op
 page one of three pages         50000   42.30 µs/op
 page one of two pages           50000   42.33 µs/op
 page three of three pages       50000   42.50 µs/op
 page two of three pages         20000   76.04 µs/op
```

### SPEED INCREASE BREAKDOWN

 * best case senario **61% faster**
 * second best case senario **39% faster**
 * worst case senario **35% faster**